src/requirements.txt
fastapi == 0.121.3
uvicorn == 0.38.0
python-multipart == 0.0.20
python-dotenv == 1.2.1
pydantic-settings == 2.12.0
aiofiles == 25.1.0
langchain == 1.0.8
pypdf == 6.3.0
langchain-community == 0.4.1
langchain-text-splitters == 1.0.0
motor == 3.7.1
